refactor(auth): move authOptions to lib/auth for getServerSession usage
