+++ src/app/home/page.tsx
@@ -1,7 +1,7 @@
 import SignOutBtn from '@/components/SignOutBtn'
 import { getServerSession } from 'next-auth'
 import { redirect } from 'next/navigation'
-import { authOptions } from '../api/auth/[...nextauth]/route'
+import { authOptions } from '@/lib/auth'
 
 export default async function Home() {
   const session = await getServerSession(authOptions)
